requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
alembic==1.7.7
    # via flask-migrate
aniso8601==7.0.0
    # via graphene
brython==3.10.5
    # via -r requirements.in
click==8.1.2
    # via flask
flask==2.1.1
    # via
    #   -r requirements.in
    #   flask-cors
    #   flask-graphql
    #   flask-migrate
    #   flask-sqlalchemy
flask-cors==3.0.10
    # via -r requirements.in
flask-graphql==2.0.1
    # via -r requirements.in
flask-migrate==3.1.0
    # via -r requirements.in
flask-sqlalchemy==2.5.1
    # via
    #   -r requirements.in
    #   flask-migrate
graphene==2.1.9
    # via graphene-sqlalchemy
graphene-sqlalchemy==2.3.0
    # via
    #   -r requirements.in
    #   graphene-sqlalchemy-filter
graphene-sqlalchemy-filter==1.13.0
    # via -r requirements.in
graphql-core==2.3.2
    # via
    #   flask-graphql
    #   graphene
    #   graphql-relay
    #   graphql-server-core
graphql-relay==2.0.1
    # via graphene
graphql-server-core==1.2.0
    # via flask-graphql
graphqlclient==0.2.4
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
itsdangerous==2.1.2
    # via flask
jinja2==3.1.1
    # via flask
mako==1.2.0
    # via alembic
markupsafe==2.1.1
    # via
    #   jinja2
    #   mako
promise==2.3
    # via
    #   graphene-sqlalchemy
    #   graphql-core
    #   graphql-relay
    #   graphql-server-core
python-slugify==8.0.4
    # via -r requirements.in
rx==1.6.1
    # via graphql-core
singledispatch==3.7.0
    # via graphene-sqlalchemy
six==1.16.0
    # via
    #   flask-cors
    #   graphene
    #   graphene-sqlalchemy
    #   graphql-core
    #   graphql-relay
    #   graphqlclient
    #   promise
    #   singledispatch
sqlalchemy==1.4.35
    # via
    #   alembic
    #   flask-sqlalchemy
    #   graphene-sqlalchemy
    #   graphene-sqlalchemy-filter
text-unidecode==1.3
    # via python-slugify
werkzeug==2.1.1
    # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
